Struct srml_contract::GenesisConfig

pub struct GenesisConfig<T: Trait> {
    pub transfer_fee: BalanceOf<T>,
    pub creation_fee: BalanceOf<T>,
    pub transaction_base_fee: BalanceOf<T>,
    pub transaction_byte_fee: BalanceOf<T>,
    pub contract_fee: BalanceOf<T>,
    pub call_base_fee: T::Gas,
    pub create_base_fee: T::Gas,
    pub gas_price: BalanceOf<T>,
    pub max_depth: u32,
    pub block_gas_limit: T::Gas,
    pub current_schedule: Schedule<T::Gas>,
}

Fields
transfer_fee: BalanceOf<T>

The fee required to make a transfer.

creation_fee: BalanceOf<T>

The fee required to create an account.

transaction_base_fee: BalanceOf<T>

The fee to be paid for making a transaction; the base.

transaction_byte_fee: BalanceOf<T>

The fee to be paid for making a transaction; the per-byte portion.

contract_fee: BalanceOf<T>

The fee required to create a contract instance.

call_base_fee: T::Gas

The base fee charged for calling into a contract.

create_base_fee: T::Gas

The base fee charged for creating a contract.

gas_price: BalanceOf<T>

The price of one unit of gas.

max_depth: u32

The maximum nesting level of a call/create stack.

block_gas_limit: T::Gas

The maximum amount of gas that could be expended per block.

current_schedule: Schedule<T::Gas>

Current cost schedule for contracts.