Trait srml_contract::Trait

pub trait Trait: Trait {
    type Currency: Currency<Self::AccountId>;
    type Call: Parameter + Dispatchable<Origin = Self::Origin>;
    type Event: From<Event<Self>> + Into<Self::Event>;
    type Gas: Parameter + Default + Codec + SimpleArithmetic + Bounded + Copy + As<BalanceOf<Self>> + As<u64> + As<u32>;
    type DetermineContractAddress: ContractAddressFor<CodeHash<Self>, Self::AccountId>;
    type ComputeDispatchFee: ComputeDispatchFee<Self::Call, BalanceOf<Self>>;
    type TrieIdGenerator: TrieIdGenerator<Self::AccountId>;
    type GasPayment: OnUnbalanced<NegativeImbalanceOf<Self>>;
}

Associated Types

type Currency: Currency<Self::AccountId>

type Call: Parameter + Dispatchable<Origin = Self::Origin>

The outer call dispatch type.

type Event: From<Event<Self>> + Into<Self::Event>

The overarching event type.

type Gas: Parameter + Default + Codec + SimpleArithmetic + Bounded + Copy + As<BalanceOf<Self>> + As<u64> + As<u32>

type DetermineContractAddress: ContractAddressFor<CodeHash<Self>, Self::AccountId>

A function type to get the contract address given the creator.

type ComputeDispatchFee: ComputeDispatchFee<Self::Call, BalanceOf<Self>>

A function type that computes the fee for dispatching the given Call.

It is recommended (though not required) for this function to return a fee that would be taken by executive module for regular dispatch.

type TrieIdGenerator: TrieIdGenerator<Self::AccountId>

trieid id generator

type GasPayment: OnUnbalanced<NegativeImbalanceOf<Self>>

Handler for the unbalanced reduction when making a gas payment.

Loading content...